refactor(neuron-kernel): move nonlinear custom-op impls out of common
