+++ spikingjelly/activation_based/cuda_kernel/neuron_kernel/common.py
@@ -37,9 +37,6 @@ __all__ = [
     "multistep_if_ptt",
     "multistep_lif_ptt",
     "multistep_plif_ptt",
-    "multistep_qif_ptt",
-    "multistep_izhikevich_ptt",
-    "multistep_eif_ptt",
 ]
 
 
@@ -3668,237 +3665,6 @@ def _resolve_sg_cuda_code_fun(sg):
     )
 
 
-if use_cupy_custom_op() and cupy is not None:
-
-    @torch.library.custom_op("sj::cupy_multistep_qif_forward", mutates_args=())
-    def cupy_multistep_qif_forward(
-        x_seq: torch.Tensor,
-        v_init: torch.Tensor,
-        tau: float,
-        v_threshold: float,
-        v_reset: float,
-        v_rest: float,
-        v_c: float,
-        a0: float,
-        detach_reset: bool,
-        sg_id: int,
-    ) -> tuple[torch.Tensor, torch.Tensor, torch.Tensor]:
-        sg = resolve_python_object(sg_id)
-        captured_ctx = _CapturedAutogradCtx()
-        spike_seq, v_seq = MultiStepQIFNodePTT.forward(
-            captured_ctx,
-            x_seq,
-            v_init,
-            tau,
-            v_threshold,
-            _decode_v_reset(v_reset),
-            v_rest,
-            v_c,
-            a0,
-            detach_reset,
-            _resolve_sg_cuda_code_fun(sg),
-        )
-        capture_id = _stash_capture_ctx(captured_ctx)
-        capture_token = torch.tensor(
-            capture_id, device=x_seq.device, dtype=torch.int64
-        )
-        return spike_seq, v_seq, capture_token
-
-
-    @torch.library.register_fake("sj::cupy_multistep_qif_forward")
-    def _cupy_multistep_qif_forward_fake(
-        x_seq, v_init, tau, v_threshold, v_reset, v_rest, v_c, a0, detach_reset, sg_id
-    ):
-        return (
-            x_seq.new_empty(x_seq.shape),
-            x_seq.new_empty(x_seq.shape),
-            x_seq.new_empty((), dtype=torch.int64),
-        )
-
-
-    def _setup_qif_ctx(ctx, inputs, output):
-        capture_token = output[2]
-        if capture_token.is_meta:
-            ctx.captured = None
-            return
-        ctx.captured = _take_capture_ctx(int(capture_token.item()))
-
-
-    def _qif_bw(ctx, grad_spike_seq, grad_v_seq, grad_capture_token):
-        del grad_capture_token
-        if ctx.captured is None:
-            raise RuntimeError("Missing captured context for qif backward.")
-        grads = MultiStepQIFNodePTT.backward(ctx.captured, grad_spike_seq, grad_v_seq)
-        return grads[0], grads[1], None, None, None, None, None, None, None, None
-
-
-    torch.library.register_autograd(
-        "sj::cupy_multistep_qif_forward",
-        _qif_bw,
-        setup_context=_setup_qif_ctx,
-    )
-
-    @torch.library.custom_op("sj::cupy_multistep_izhikevich_forward", mutates_args=())
-    def cupy_multistep_izhikevich_forward(
-        x_seq: torch.Tensor,
-        v_init: torch.Tensor,
-        w_init: torch.Tensor,
-        tau: float,
-        v_threshold: float,
-        v_reset: float,
-        v_rest: float,
-        a: float,
-        b: float,
-        tau_w: float,
-        v_c: float,
-        a0: float,
-        detach_reset: bool,
-        sg_id: int,
-    ) -> tuple[torch.Tensor, torch.Tensor, torch.Tensor, torch.Tensor]:
-        sg = resolve_python_object(sg_id)
-        captured_ctx = _CapturedAutogradCtx()
-        spike_seq, v_seq, w_seq = MultiStepIzhikevichNodePTT.forward(
-            captured_ctx,
-            x_seq,
-            v_init,
-            w_init,
-            tau,
-            v_threshold,
-            _decode_v_reset(v_reset),
-            v_rest,
-            a,
-            b,
-            tau_w,
-            v_c,
-            a0,
-            detach_reset,
-            _resolve_sg_cuda_code_fun(sg),
-        )
-        capture_id = _stash_capture_ctx(captured_ctx)
-        capture_token = torch.tensor(
-            capture_id, device=x_seq.device, dtype=torch.int64
-        )
-        return spike_seq, v_seq, w_seq, capture_token
-
-
-    @torch.library.register_fake("sj::cupy_multistep_izhikevich_forward")
-    def _cupy_multistep_izhikevich_forward_fake(*args):
-        x_seq = args[0]
-        return (
-            x_seq.new_empty(x_seq.shape),
-            x_seq.new_empty(x_seq.shape),
-            x_seq.new_empty(x_seq.shape),
-            x_seq.new_empty((), dtype=torch.int64),
-        )
-
-
-    def _setup_iz_ctx(ctx, inputs, output):
-        capture_token = output[3]
-        if capture_token.is_meta:
-            ctx.captured = None
-            return
-        ctx.captured = _take_capture_ctx(int(capture_token.item()))
-
-
-    def _iz_bw(ctx, grad_spike_seq, grad_v_seq, grad_w_seq, grad_capture_token):
-        del grad_capture_token
-        if ctx.captured is None:
-            raise RuntimeError("Missing captured context for izhikevich backward.")
-        grads = MultiStepIzhikevichNodePTT.backward(
-            ctx.captured, grad_spike_seq, grad_v_seq, grad_w_seq
-        )
-        return (
-            grads[0],
-            grads[1],
-            grads[2],
-            None,
-            None,
-            None,
-            None,
-            None,
-            None,
-            None,
-            None,
-            None,
-            None,
-            None,
-        )
-
-
-    torch.library.register_autograd(
-        "sj::cupy_multistep_izhikevich_forward",
-        _iz_bw,
-        setup_context=_setup_iz_ctx,
-    )
-
-    @torch.library.custom_op("sj::cupy_multistep_eif_forward", mutates_args=())
-    def cupy_multistep_eif_forward(
-        x_seq: torch.Tensor,
-        v_init: torch.Tensor,
-        tau: float,
-        v_threshold: float,
-        v_reset: float,
-        v_rest: float,
-        theta_rh: float,
-        delta_T: float,
-        detach_reset: bool,
-        sg_id: int,
-    ) -> tuple[torch.Tensor, torch.Tensor, torch.Tensor]:
-        sg = resolve_python_object(sg_id)
-        captured_ctx = _CapturedAutogradCtx()
-        spike_seq, v_seq = MultiStepEIFNodePTT.forward(
-            captured_ctx,
-            x_seq,
-            v_init,
-            tau,
-            v_threshold,
-            _decode_v_reset(v_reset),
-            v_rest,
-            theta_rh,
-            delta_T,
-            detach_reset,
-            _resolve_sg_cuda_code_fun(sg),
-        )
-        capture_id = _stash_capture_ctx(captured_ctx)
-        capture_token = torch.tensor(
-            capture_id, device=x_seq.device, dtype=torch.int64
-        )
-        return spike_seq, v_seq, capture_token
-
-
-    @torch.library.register_fake("sj::cupy_multistep_eif_forward")
-    def _cupy_multistep_eif_forward_fake(*args):
-        x_seq = args[0]
-        return (
-            x_seq.new_empty(x_seq.shape),
-            x_seq.new_empty(x_seq.shape),
-            x_seq.new_empty((), dtype=torch.int64),
-        )
-
-
-    def _setup_eif_ctx(ctx, inputs, output):
-        capture_token = output[2]
-        if capture_token.is_meta:
-            ctx.captured = None
-            return
-        ctx.captured = _take_capture_ctx(int(capture_token.item()))
-
-
-    def _eif_bw(ctx, grad_spike_seq, grad_v_seq, grad_capture_token):
-        del grad_capture_token
-        if ctx.captured is None:
-            raise RuntimeError("Missing captured context for eif backward.")
-        grads = MultiStepEIFNodePTT.backward(ctx.captured, grad_spike_seq, grad_v_seq)
-        return grads[0], grads[1], None, None, None, None, None, None, None, None
-
-
-    torch.library.register_autograd(
-        "sj::cupy_multistep_eif_forward",
-        _eif_bw,
-        setup_context=_setup_eif_ctx,
-    )
-
-
 def multistep_if_ptt(
     x_seq, v_init, v_threshold, v_reset, detach_reset, surrogate_function
 ):
@@ -3935,92 +3701,6 @@ def multistep_plif_ptt(
         detach_reset,
         _resolve_sg_cuda_code_fun(surrogate_function),
     )
-
-
-def multistep_qif_ptt(
-    x_seq, v_init, tau, v_threshold, v_reset, v_rest, v_c, a0, detach_reset, surrogate_function
-):
-    sg_id = _sg_obj_id(surrogate_function)
-    v_reset_value = float("nan") if v_reset is None else float(v_reset)
-    spike_seq, v_seq, _ = cupy_multistep_qif_forward(
-        x_seq,
-        v_init,
-        tau,
-        v_threshold,
-        v_reset_value,
-        v_rest,
-        v_c,
-        a0,
-        detach_reset,
-        sg_id,
-    )
-    return spike_seq, v_seq
-
-
-def multistep_izhikevich_ptt(
-    x_seq,
-    v_init,
-    w_init,
-    tau,
-    v_threshold,
-    v_reset,
-    v_rest,
-    a,
-    b,
-    tau_w,
-    v_c,
-    a0,
-    detach_reset,
-    surrogate_function,
-):
-    sg_id = _sg_obj_id(surrogate_function)
-    v_reset_value = float("nan") if v_reset is None else float(v_reset)
-    spike_seq, v_seq, w_seq, _ = cupy_multistep_izhikevich_forward(
-        x_seq,
-        v_init,
-        w_init,
-        tau,
-        v_threshold,
-        v_reset_value,
-        v_rest,
-        a,
-        b,
-        tau_w,
-        v_c,
-        a0,
-        detach_reset,
-        sg_id,
-    )
-    return spike_seq, v_seq, w_seq
-
-
-def multistep_eif_ptt(
-    x_seq,
-    v_init,
-    tau,
-    v_threshold,
-    v_reset,
-    v_rest,
-    theta_rh,
-    delta_T,
-    detach_reset,
-    surrogate_function,
-):
-    sg_id = _sg_obj_id(surrogate_function)
-    v_reset_value = float("nan") if v_reset is None else float(v_reset)
-    spike_seq, v_seq, _ = cupy_multistep_eif_forward(
-        x_seq,
-        v_init,
-        tau,
-        v_threshold,
-        v_reset_value,
-        v_rest,
-        theta_rh,
-        delta_T,
-        detach_reset,
-        sg_id,
-    )
-    return spike_seq, v_seq
 
 
 def save_cuda_codes(
